merging/reducing util files
