--- src/data-mgt/node/utils/jsonify.js
@@ -1,7 +1,0 @@
-const jsonParse = (dataFromDatabase) => {
-  let jsonData = JSON.stringify(dataFromDatabase);
-  let parsedMap = JSON.parse(jsonData);
-  return parsedMap;
-};
-
-module.exports = jsonParse;
